LIBASPACE-261. umd_location_ead_export proof-of-concept
Added a plugin to export the first "Location" field of a Top Container
into the "label" attribute of "container" tags.

To use this plugin, edit the "common/config/config.rb" file, adding
the following line as the the third line:

```
AppConfig[:plugins] << 'umd_location_ead_export'
```

Also, when loading the sample data, only the "0057.LIT_20171121_135754_UTC__ead.xml" file needs to be loaded for the following steps.

https://issues.umd.edu/browse/LIBASPACE-261